+++ CHANGES.md
@@ -8,6 +8,7 @@ Changes in 8.1
 - Remove deprecated DBSeqConvertSerialName() function in database.inc.php
 - Fix Conflict Warning displayed twice in Courses.php
 - Fix PHP Notice Undefined index in miscExport.fnc.php
+- Fix SQL error when Student / Staff ID is hacked / not an integer in URL in Current.php
 
 Changes in 8.0.4
 ----------------